[Silver IV] Title: 피하자, Time: 608 ms, Memory: 143644 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/25379. 피하자/피하자.java
@@ -31,11 +31,11 @@ public class Main {
             return;
         }
 
-        int cnt = 0;
+        long cnt = 0;
 
         // 짝수를 왼쪽에 정렬하는 방법
         int start = odd.get(0);
-        int temp = 0;
+        long temp = 0;
         for (int i = 0 ; i < even.size() ; i++) {
             if (even.get(i) < start) continue;
